IOT-350: Move resource handlers from webservice to their respective modules. IOT-84: Make ParserBolt use DataSourceDto instead of DataSource
