+++ streams/service/src/main/java/com/hortonworks/iotas/streams/service/DataSourceFacade.java
@@ -1,4 +1,4 @@
-package com.hortonworks.iotas.webservice.catalog;
+package com.hortonworks.iotas.streams.service;
 
 import com.hortonworks.iotas.catalog.DataFeed;
 import com.hortonworks.iotas.catalog.DataSource;
@@ -12,7 +12,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.Map;
 import java.util.HashMap;
 import java.util.List;
